Cleaned up warnings from Visual C++ 6.0 compile by adding casts.

+++ trunk/bacnet-stack/demo/handler/s_rp.c
@@ -89,7 +89,7 @@ uint8_t Send_Read_Property_Request(uint32_t device_id,  /* destination device */
            max_apdu in the address binding table. */
         if ((unsigned) pdu_len < max_apdu) {
             tsm_set_confirmed_unsegmented_transaction(invoke_id, &dest,
-                &npdu_data, &Handler_Transmit_Buffer[0], pdu_len);
+                &npdu_data, &Handler_Transmit_Buffer[0], (uint16_t)pdu_len);
             bytes_sent =
                 datalink_send_pdu(&dest, &npdu_data,
                 &Handler_Transmit_Buffer[0], pdu_len);
